Implement Primary.xml file parser
Parsing the rpm's Primary.xml packages metadata file using
pulldom xml parser as a memory efficient parsing library.

Note that some attributes in the returned parsed object are
faked, and maybe filled in elsewhere.

The faking of some of the data is done because some
attributes are required by the importer service.