Fix importing updates
In the previous approach, we tried to split the content of the
'available_packages' dict into small pieces, each peace contains
the packages of one batch at a time, in order to save memory by
not storing all the packages' info (in the form of { name-epoch-version-release-arch : 1 })
in memory at once.

However, this was wrong, because the UpdatesImporter needed to
verify on all the packages each time, and not a portion of them,
so it gave wrong results.

So, we'll have to store the whole packages' dict in memory,
for a short period of time, to be used by the UpdatesImporter.

And also, in the previous approach, we tried to run the
UpdatesImporter after importing each batch of packages,
which might not also be correct, because the UpdatesImporter
is fetching the imported packages information from the db
and checking on some conditions on them...